requirements.txt
antlr4-python3-runtime==4.13.2
appnope==0.1.4
asttokens==2.4.1
beautifulsoup4==4.12.3
betterproto==2.0.0b6
black==24.3.0
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
grpclib==0.4.7
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
thoughtspot_tml==2.1.1
tomli==2.0.1
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
zipp==3.18.1
SQLAlchemy==2.0.35
snowflake-connector-python==3.12.2
snowflake-sqlalchemy==1.6.1
sortedcontainers==2.4.0
cryptography==43.0.1
python-dotenv==1.0.1
snowflake-snowpark-python==1.22.1
